No deja eliminar 0

+++ Eliminacion/FEConductores.php
@@ -55,6 +55,10 @@
 		$datoBorrar=$_POST['datoBorrar'];
 		print("Identificador: ".$datoBorrar);
 		print("<br>");
+		if ($datoBorrar==0) {
+			print("Error, no se puede borrar este dato");
+			exit();
+		}
 		include("conexion.php");
 		$Con= Conectar();
 		$SQL = "DELETE FROM conductores WHERE rfc = '$datoBorrar';"; //DE QUE TABLA Y CONDICION
