+++ Eliminacion/FEPropietarios.php
@@ -57,6 +57,10 @@
 		$datoBorrar=$_POST['datoBorrar'];
 		print("Identificador: ".$datoBorrar);
 		print("<br>");
+		if ($datoBorrar==0) {
+			print("Error, no se puede borrar este dato");
+			exit();
+		}
 		include("conexion.php");
 		$Con= Conectar();
 		$SQL = "DELETE FROM propietarios WHERE rfc = '$datoBorrar';"; //DE QUE TABLA Y CONDICION
